+++ iotdb-core/confignode/src/main/java/org/apache/iotdb/confignode/manager/pipe/transfer/connector/payload/request/PipeTransferConfigNodeHandshakeV1Req.java
@@ -20,7 +20,6 @@
 package org.apache.iotdb.confignode.manager.pipe.transfer.connector.payload.request;
 
 import org.apache.iotdb.commons.pipe.connector.payload.thrift.request.PipeRequestType;
-import org.apache.iotdb.commons.pipe.connector.payload.thrift.request.PipeTransferHandshakeV1Req;
 import org.apache.iotdb.service.rpc.thrift.TPipeTransferReq;
 
 import java.io.IOException;
